`Torchapp` depends on `torch`, however, `torch` is a very heavy import. This significantly slows down the responsiveness of `torchapp` CLI apps. In this PR I've made aggressive use of lazy importing wherever I can to prevent `torch` from being imported until it is required. This also means you need to slighting change the way you write apps that inherit from `torchapp` so that they also use lazy importing. I've updated a few of the examples to use lazy importing to demonstrate this. 

**Lazy importing results in about a 20x CLI speed up.**

---

Before lazy importing: importing `lightning`, `torch` and `pandas` using >90% of the time just to call `--help`. 

```bash
❯ hyperfine "python torchapp/examples/logistic_regression.py --help"
Benchmark 1: python torchapp/examples/logistic_regression.py --help
  Time (mean ± σ):      3.319 s ±  0.406 s    [User: 3.643 s, System: 1.375 s]
  Range (min … max):    3.072 s …  4.444 s    10 runs
``` 

![image](https://github.com/user-attachments/assets/efc4fc63-0b63-4bcd-a91a-8f7194f6a590)

After lazy importing: `lightning`, `torch` and `pandas` are not imported at all.

```bash
❯ hyperfine "python torchapp/examples/logistic_regression.py --help"                                        
Benchmark 1: python torchapp/examples/logistic_regression.py --help
  Time (mean ± σ):     182.0 ms ±  46.0 ms    [User: 125.3 ms, System: 33.0 ms]
  Range (min … max):   155.2 ms … 311.1 ms    10 runs
```

![image](https://github.com/user-attachments/assets/b94b9b89-662b-459f-ad7f-f09c3e88a8bd)